fcgi: make it possible to use the installed version of flup instead of builtin version.
The environment varialbe TRAC_USE_FLUP needs to be set to `1` in order to force the use of flup.
If `flup` is used, we need to insert some WSGI middleware in order to deal with URL encoding.

Related to #9111.


git-svn-id: http://trac.edgewall.org/intertrac/log:/branches/0.12-stable@10004 af82e41b-90c4-0310-8c96-b1721e28e2e2
