+++ trac/web/standalone.py
@@ -23,7 +23,6 @@ import pkg_resources
 import os
 import sys
 from SocketServer import ThreadingMixIn
-import urllib
 
 from trac import __version__ as VERSION
 from trac.util import autoreload, daemon
@@ -68,16 +67,6 @@ class BasePathMiddleware(object):
         path = environ['SCRIPT_NAME'] + environ.get('PATH_INFO', '')
         environ['PATH_INFO'] = path[len(self.base_path):]
         environ['SCRIPT_NAME'] = self.base_path
-        return self.application(environ, start_response)
-
-
-class FlupMiddleware(object):
-
-    def __init__(self, application):
-        self.application = application
-
-    def __call__(self, environ, start_response):
-        environ['PATH_INFO'] = urllib.unquote(environ.get('PATH_INFO', ''))
         return self.application(environ, start_response)
 
 
@@ -272,6 +261,7 @@ def main():
                                     None, None, ['']).WSGIServer
             flup_app = wsgi_app
             if options.unquote:
+                from trac.web.fcgi_frontend import FlupMiddleware
                 flup_app = FlupMiddleware(flup_app)
             ret = server_cls(flup_app, bindAddress=server_address).run()
             sys.exit(ret and 42 or 0) # if SIGHUP exit with status 42
